Sửa cấu trúc file, tạo trang chủ, đổi navbar, chỉnh hiệu ứng cho việc chuyển tab, thêm footer
