+++ package.json
@@ -10,7 +10,9 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "bootstrap": "^5.3.3",
+    "bourbon": "^7.3.0",
     "clsx": "^2.1.1",
+    "compass": "^0.1.1",
     "convex": "^1.15.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
