+++ src/components/App.js
@@ -1,16 +1,17 @@
 import { Route, Routes } from 'react-router';
 import 'bootstrap/dist/css/bootstrap.min.css';
 
-import PredictPage from './PredictPage/PredictPage';
-import Login from './Login/Login';
+import { publicRoutes } from '../routes';
 
 function App() {
   return (
     <>
       <Routes>
-        <Route path='/' element={<Login />}/>
-      <Route path='/predict-page' element={<PredictPage />} />
-    </Routes >
+        {publicRoutes.map((route, index) => {
+          const Page = route.component
+          return <Route key={index} path={route.path} element={<Page />}/>
+        })}
+      </Routes >
     </>
   );
 }
