+++ src/components/App.module.css
@@ -1,0 +1,7 @@
+@tailwind base;
+
+@layer base {
+    body {
+        @apply font-roboto;
+    }
+}
