--- src/components/PredictPage/Header/Header.module.css
@@ -1,23 +1,0 @@
-.root {
-  border-bottom: 1px solid rgb(148, 148, 148);
-  margin-bottom: 10px;
-  background-color: #dbeafe;
-  position: relative;
-  box-shadow: 3px 3px 1px #8a8a8a;
-  border-radius: 7px;
-}
-
-.logo {
-  width: 150px;
-  position: absolute;
-}
-
-.title {
-  color: #0e4051;
-}
-
-@media (max-width: 770px) {
-    .logo {
-      display: none;
-    }
-}
