--- src/components/PredictPage/SidePanel/SidePanel.module.css
@@ -1,60 +1,0 @@
-.root {
-  margin-left: 10px;
-}
-.active {
-  color: #000000 !important; 
-  background-color: #c7c7c7;
-}
-.item {
-  display: flex;
-  font-size: 18px;
-  font-family: ui-sans-serif,system-ui,sans-serif;
-  text-align: center;
-  align-items: center;
-  color: #5e5e5e !important;
-  font-size: 12px;
-  font-weight: bold;
-  width: 90px;
-  height: 90px;
-  border-radius: 20px !important;
-  margin-top: 20px;
-  text-decoration: none;
-}
-
-
-.item:hover {
-  border: none !important;
-  background-color: #c7c7c7;
-  color: #000000;
-  font-size: 13px;
-}
-
-@media (max-width: 993px) {
-  .parent {
-    /* width: 100px;
-    background-color: #fff;
-    border: 1px solid #000; */
-    position: fixed;
-    bottom: 0px;
-    width: 100%;
-    height: 80px;
-    background-color: #dbeafe;
-    border-top: 1px solid #d0d0d0;
-    display: flex;
-    /* border: 1px solid #000; */
-    margin-left: -12px;
-  }
-
-  .item {
-    font-size: 11px !important;
-    width: 70px;
-    height: 70px;
-  }
-
-  .root {
-    margin-left: 0px;
-    margin-top: -15px;
-    justify-content: center !important;
-    display: flex;
-  }
-}
