+++ src/components/SidePanel/SidePanel.module.css
@@ -1,0 +1,54 @@
+.root {
+  margin-left: 10px;
+}
+.active {
+  color: #000000 !important; 
+  background-color: #c7c7c7;
+}
+
+.parent {
+  display: flex;
+  justify-content: space-evenly;
+  flex-wrap: wrap;
+  padding: 0;
+}
+.frameNav {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: #63b6e7;
+  width: 22vh;
+  border-radius: 20px;
+  height: 10vh;
+  margin-top: 10px;
+  margin-bottom: 10px;
+  list-style-type: none !important;
+  cursor: pointer;
+}
+.item {
+  margin-left: -3px;
+  background-color: inherit;
+  text-decoration: none;
+  color: #000000;
+  text-align: center;
+}
+
+
+@media (max-width: 993px) {
+  /* .parent {
+    position: fixed;
+    bottom: 0px;
+    width: 100%;
+    height: 80px;
+    background-color: #dbeafe;
+    border-top: 1px solid #d0d0d0;
+    display: flex;
+    margin-left: -12px;
+  } */
+
+  .item {
+    font-size: 11px !important;
+    width: 70px;
+    height: 70px;
+  }
+}
