+++ src/pages/PredictPage/PredictBaseOnQuesstion/PredictBaseOnQesstion.module.css
