+++ src/pages/PredictPage/PredictFinalScore/PredictFinalScore.module.css
@@ -1,0 +1,115 @@
+.title {
+    height: 90px;
+    background-color: #75d1ff;
+    color: #FFF;
+    font-family: 'Segoe UI';
+    display: flex;
+    align-items: center;
+    border-radius: 20px;
+
+}
+
+.btnAnalysis {
+    width: 120px;
+    height: 40px;
+    background-color: #f69f7f;
+    border: none;
+    /* align-items: center !important; */
+    margin: 0 auto;
+    vertical-align: middle !important;
+    margin-top: -2px;
+}
+
+.btnAnalysis:hover {
+    background-color: #fd7949;
+}
+
+.btnAnalysis:active {
+    background-color: #f69f7f !important;
+    transform: scale(0.95);
+}
+
+.btnPredict {
+    width: 120px;
+    height: 40px;
+    background-color: #4dc529;
+    border: none;
+    /* align-items: center !important; */
+    margin: 0 auto;
+    vertical-align: middle !important;
+    margin-top: -2px;
+    margin-left: 10px;
+    
+}
+
+.btnPredict:hover {
+    background-color: #34b40e;
+}
+
+.btnPredict:active {
+    background-color: #4dc529 !important;
+    transform: scale(0.95);
+}
+
+.frameTab {
+    margin-left: 30px;
+    margin-right: 10px;
+}
+
+.borderChart {
+    border: 1px solid #929292;
+    height: 500px;
+    margin-left: 10px;
+    width: 98%;
+    border-radius: 10px;
+    box-shadow: 5px 4px #888888;
+}
+
+.label {
+    margin-top: 15px;
+    /* padding-right: 60px; */
+}
+
+.frameInput {
+    padding-right: 20px !important;
+}
+
+.input {
+    margin-top: 9px;
+}
+
+@media (max-width: 993px) {
+    .label {
+        text-align: start !important;
+        margin-top: 0;
+        margin-left: 4px !important;
+    }
+
+    .footerTemp {
+        height: 120px;
+        width: 100%;
+    }
+
+    .input {
+        margin-top: 0px;
+        margin-bottom: 10px;
+    }
+
+    .btnPredict {
+        margin-left: 0;
+        margin-top: 8px;
+    }
+
+    .predictText {
+        margin-left: -12px !important;
+    }
+
+    .frameInput {
+        padding-right: 30px !important;
+    }
+    
+}
+
+.predictText {
+    margin-left: -6px;
+}
