+++ src/pages/PredictPage/PredictFinalScore/PredictTextFinalScore/PredictTextFinalScore.module.css
@@ -1,0 +1,13 @@
+.label {
+    margin-top: 15px;
+    /* padding-right: 60px; */
+}
+
+@media (max-width: 993px) {
+    .label {
+        text-align: start !important;
+        margin-top: 0;
+        margin-bottom: -10px;
+        margin-left: 4px !important;
+    }
+}
