+++ src/pages/PredictPage/PredictInlab/LabItems/Lab.js
