+++ src/pages/PredictPage/Predictpage.module.css
@@ -1,0 +1,24 @@
+/* .root {
+    position: relative;
+    z-index: 0; 
+    width: 100vw; 
+    height: 100vh;
+}
+
+.root::before {
+    content: "";
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-image: url('../../images/nen-iuh.jpg');
+    background-size: cover;
+    background-position:top;
+    opacity: 0.2;
+    z-index: -1;
+} */
+
+.footer {
+    background-color: #dbeafe; 
+}
