+++ src/store/constants.js
@@ -6,3 +6,4 @@ export const SET_RESET = 'set_reset' //nút reset
 //Bai toan 3
 export const SET_ACTIVE_TAB_LEFT = 'set_active_tab_left' // thanh tabs bên trái
 export const SET_STUDENT_ID = 'set_student_id' //mã số sinh viên
+export const SET_PREDICTED_VALUE_FINAL = 'set_predicted_value_final'
